V. 0.2.0
Two-move deep analysis. Checking if there is a possibility for white to take a pawn after considered brown move and if there is - do not consider this game node as a possible move, which prevents giving up pawns stupidly.

Also simplfied code for algorithm. Plus deciding when to finish game and announce a victor (and add a game to statistics and history later).